+++ LegendWeathers/Patches/BloodMoonPatches.cs
@@ -1,6 +1,7 @@
 ﻿using HarmonyLib;
 using LegendWeathers.Weathers;
 using LegendWeathers.WeatherSkyEffects;
+using Unity.Netcode;
 
 namespace LegendWeathers.Patches
 {
@@ -35,6 +36,49 @@ namespace LegendWeathers.Patches
             if ((!destroy || !__instance.enemyType.canBeDestroyed) && __instance.enemyType.canDie)
             {
                 BloodMoonWeather.BloodMoonEffectReference.WorldObject?.GetComponent<BloodMoonWeather>()?.GetBloodMoonManager()?.ResurrectEnemy(__instance);
+            }
+
+        }
+    }
+
+
+    [HarmonyPatch(typeof(NutcrackerEnemyAI))]
+    internal class BloodMoonNutcrackerPatch
+    {
+        [HarmonyPrefix]
+        [HarmonyPatch("SpawnShotgunShells")]
+        public static bool SpawnShotgunShellsPatch(NutcrackerEnemyAI __instance)
+        {
+            return !(Plugin.config.bloodMoonWeather.Value && BloodMoonWeather.BloodMoonEffectReference != null && BloodMoonWeather.BloodMoonEffectReference.EffectEnabled
+                && __instance != null && __instance.enemyType != null);
+        }
+
+        [HarmonyPrefix]
+        [HarmonyPatch("DropGun")]
+        public static bool DropGunPatch(NutcrackerEnemyAI __instance)
+        {
+            if (!Plugin.config.bloodMoonWeather.Value || BloodMoonWeather.BloodMoonEffectReference == null || !BloodMoonWeather.BloodMoonEffectReference.EffectEnabled
+                || __instance == null || __instance.enemyType == null || __instance.gun == null)
+            {
+                return true;
+            }
+            if (!BloodMoonManager.hasFirstNutcrackerRespawned)
+            {
+                __instance.gun.shellsLoaded = 2;
+                BloodMoonManager.hasFirstNutcrackerRespawned = true;
+                return true;
+            }
+            else
+            {
+                if (__instance.IsServer)
+                {
+                    var gunNetwork = __instance.gun.gameObject.GetComponent<NetworkObject>();
+                    if (gunNetwork != null && gunNetwork.IsSpawned)
+                    {
+                        gunNetwork.Despawn();
+                    }
+                }
+                return false;
             }
         }
     }